requirements.txt
ipykernel

load_dotenv
pydantic

langchain
langgraph
langchain_google_genai